tests/wpt/tests/html/canvas/offscreen/layers/2d.layer.malformed-operations.transferToImageBitmap.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.malformed-operations.transferToImageBitmap</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.malformed-operations.transferToImageBitmap</h1>
<p class="desc">Check that exceptions are thrown for operations that are malformed while layers are open.</p>


<script>
var t = async_test("Check that exceptions are thrown for operations that are malformed while layers are open.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(200, 200);
  var ctx = canvas.getContext('2d');

  // Shouldn't throw on its own.
  canvas.transferToImageBitmap();
  // Make sure the exception isn't caused by calling the function twice.
  canvas.transferToImageBitmap();
  // Calling again inside a layer should throw.
  ctx.beginLayer();
  assert_throws_dom("InvalidStateError",
                    () => canvas.transferToImageBitmap());
  t.done();

});
</script>
